+++ src/kernels/update_velocities.cu
@@ -22,9 +22,9 @@ __device__
 {
     for (fiberuint quadrature_index_i = 0; quadrature_index_i < TOTAL_NUMBER_OF_QUADRATURE_POINTS; ++quadrature_index_i)
     {
-        GF[quadrature_index_i + 0 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
-        GF[quadrature_index_i + 1 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
-        GF[quadrature_index_i + 2 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0;
+        GF[quadrature_index_i + 0 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
+        GF[quadrature_index_i + 1 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
+        GF[quadrature_index_i + 2 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] = 0.0f;
 
         fiberfloat4 position_on_fiber_i;
         position_on_fiber_i.x = position_i.x + quadrature_points[quadrature_index_i] * orientation_i.x;
@@ -44,42 +44,42 @@ __device__
             difference.y = position_on_fiber_i.y - position_on_fiber_j.y;
             difference.z = position_on_fiber_i.z - position_on_fiber_j.z;
 
-            const fiberfloat distance = sqrt(difference.x * difference.x + difference.y * difference.y + difference.z * difference.z);
+            const fiberfloat distance = sqrtf(difference.x * difference.x + difference.y * difference.y + difference.z * difference.z);
 
             // equation 10
             // Note:    The outer product of a vector with itself is always a symmetric matrix
             //          so to save computation we only compute the upper triangle.
             // TODO calculation can be optimized (i.e. not dividing by distance, simpifing etc.)
-            const fiberfloat K11 = 1.0 / distance
-                                   + (1.0 / distance) * (difference.x / distance) * (difference.x / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS * ((1.0 / (distance * distance * distance))
-                                           - (3.0 / (distance * distance * distance)) * ((difference.x / distance) * (difference.x / distance)));
-            const fiberfloat K22 = 1.0 / distance
-                                   + (1.0 / distance) * (difference.y / distance) * (difference.y / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS * ((1.0 / (distance * distance * distance))
-                                           - (3.0 / (distance * distance * distance)) * ((difference.y / distance) * (difference.y / distance)));
-            const fiberfloat K33 = 1.0 / distance
-                                   + (1.0 / distance) * (difference.z / distance) * (difference.z / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS * ((1.0 / (distance * distance * distance))
-                                           - (3.0 / (distance * distance * distance)) * ((difference.z / distance) * (difference.z / distance)));
-            const fiberfloat K12 = (1.0 / distance) * (difference.x / distance) * (difference.y / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS
-                                   * (-3.0 / (distance * distance * distance)) * (difference.x / distance) * (difference.y / distance);
+            const fiberfloat K11 = 1.0f / distance
+                                   + (1.0f / distance) * (difference.x / distance) * (difference.x / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS * ((1.0f / (distance * distance * distance))
+                                           - (3.0f / (distance * distance * distance)) * ((difference.x / distance) * (difference.x / distance)));
+            const fiberfloat K22 = 1.0f / distance
+                                   + (1.0f / distance) * (difference.y / distance) * (difference.y / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS * ((1.0f / (distance * distance * distance))
+                                           - (3.0f / (distance * distance * distance)) * ((difference.y / distance) * (difference.y / distance)));
+            const fiberfloat K33 = 1.0f / distance
+                                   + (1.0f / distance) * (difference.z / distance) * (difference.z / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS * ((1.0f / (distance * distance * distance))
+                                           - (3.0f / (distance * distance * distance)) * ((difference.z / distance) * (difference.z / distance)));
+            const fiberfloat K12 = (1.0f / distance) * (difference.x / distance) * (difference.y / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS
+                                   * (-3.0f / (distance * distance * distance)) * (difference.x / distance) * (difference.y / distance);
 
-            const fiberfloat K13 = (1.0 / distance) * (difference.x / distance) * (difference.z / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS
-                                   * (-3.0 / (distance * distance * distance)) * (difference.x / distance) * (difference.z / distance);
+            const fiberfloat K13 = (1.0f / distance) * (difference.x / distance) * (difference.z / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS
+                                   * (-3.0f / (distance * distance * distance)) * (difference.x / distance) * (difference.z / distance);
 
-            const fiberfloat K23 = (1.0 / distance) * (difference.y / distance) * (difference.z / distance)
-                                   + 2.0 * SLENDERNESS * SLENDERNESS
-                                   * (-3.0 / (distance * distance * distance)) * (difference.y / distance) * (difference.z / distance);
+            const fiberfloat K23 = (1.0f / distance) * (difference.y / distance) * (difference.z / distance)
+                                   + 2.0f * SLENDERNESS * SLENDERNESS
+                                   * (-3.0f / (distance * distance * distance)) * (difference.y / distance) * (difference.z / distance);
 
             const fiberfloat quadrature_weight = quadrature_weights[quadrature_index_j];
 
             fiberfloat4 force_on_fiber_j;
-            force_on_fiber_j.x = 0.5 * external_force.x;
-            force_on_fiber_j.y = 0.5 * external_force.y;
-            force_on_fiber_j.z = 0.5 * external_force.z;
+            force_on_fiber_j.x = 0.5f * external_force.x;
+            force_on_fiber_j.y = 0.5f * external_force.y;
+            force_on_fiber_j.z = 0.5f * external_force.z;
 
             for (fiberuint force_index = 0; force_index < NUMBER_OF_TERMS_IN_FORCE_EXPANSION; ++force_index)
             {
@@ -119,7 +119,7 @@ __global__ void update_velocities(
 
     if (i >= NUMBER_OF_FIBERS) return;
 
-    const fiberfloat c  = log(SLENDERNESS * SLENDERNESS * M_E);
+    const fiberfloat c  = logf(SLENDERNESS * SLENDERNESS * M_E);
     const fiberfloat d  = -c;
 
     const fiberfloat4 position_i = positions[i];
@@ -127,22 +127,22 @@ __global__ void update_velocities(
 
     // @TODO Constant external force
     fiberfloat4 external_force;
-    external_force.x = 0.0;
-    external_force.y = 0.0;
-    external_force.z = -1.0;
+    external_force.x = 0.0f;
+    external_force.y = 0.0f;
+    external_force.z = -1.0f;
 
     fiberfloat4 oriented_force;
     oriented_force.x = orientation_i.x * orientation_i.x * external_force.x + orientation_i.x * orientation_i.y * external_force.y + orientation_i.x * orientation_i.z * external_force.z;
     oriented_force.y = orientation_i.x * orientation_i.y * external_force.x + orientation_i.y * orientation_i.y * external_force.y + orientation_i.y * orientation_i.z * external_force.z;
     oriented_force.z = orientation_i.x * orientation_i.z * external_force.x + orientation_i.y * orientation_i.z * external_force.y + orientation_i.z * orientation_i.z * external_force.z;
 
-    translational_velocities[i].x = 0.5 * ((d + 2) * external_force.x + (d - 2) * oriented_force.x);
-    translational_velocities[i].y = 0.5 * ((d + 2) * external_force.y + (d - 2) * oriented_force.y);
-    translational_velocities[i].z = 0.5 * ((d + 2) * external_force.z + (d - 2) * oriented_force.z);
+    translational_velocities[i].x = 0.5f * ((d + 2.0f) * external_force.x + (d - 2.0f) * oriented_force.x);
+    translational_velocities[i].y = 0.5f * ((d + 2.0f) * external_force.y + (d - 2.0f) * oriented_force.y);
+    translational_velocities[i].z = 0.5f * ((d + 2.0f) * external_force.z + (d - 2.0f) * oriented_force.z);
 
-    rotational_velocities[i].x = 0.0;
-    rotational_velocities[i].y = 0.0;
-    rotational_velocities[i].z = 0.0;
+    rotational_velocities[i].x = 0.0f;
+    rotational_velocities[i].y = 0.0f;
+    rotational_velocities[i].z = 0.0f;
 
     for (fiberuint j = 0; j < NUMBER_OF_FIBERS; ++j)
     {
@@ -154,13 +154,13 @@ __global__ void update_velocities(
         fiberfloat GF[24 * 3];
         compute_GV(j, position_i, orientation_i, position_j, orientation_j, coefficients, external_force, quadrature_points, quadrature_weights, legendre_polynomials, GF, SLENDERNESS, NUMBER_OF_TERMS_IN_FORCE_EXPANSION, TOTAL_NUMBER_OF_QUADRATURE_POINTS);
 
-        fiberfloat TF1A0 = 0.0;
-        fiberfloat TF2A0 = 0.0;
-        fiberfloat TF3A0 = 0.0;
+        fiberfloat TF1A0 = 0.0f;
+        fiberfloat TF2A0 = 0.0f;
+        fiberfloat TF3A0 = 0.0f;
 
-        fiberfloat TF1A1 = 0.0;
-        fiberfloat TF2A1 = 0.0;
-        fiberfloat TF3A1 = 0.0;
+        fiberfloat TF1A1 = 0.0f;
+        fiberfloat TF2A1 = 0.0f;
+        fiberfloat TF3A1 = 0.0f;
 
         for (fiberuint quadrature_index_i = 0; quadrature_index_i < TOTAL_NUMBER_OF_QUADRATURE_POINTS; ++quadrature_index_i)
         {
@@ -176,13 +176,13 @@ __global__ void update_velocities(
             TF3A1 += quadrature_weight * GF[quadrature_index_i + 2 * TOTAL_NUMBER_OF_QUADRATURE_POINTS] * legendre_polynomial;
         }
 
-        translational_velocities[i].x += 0.5 * TF1A0;
-        translational_velocities[i].y += 0.5 * TF2A0;
-        translational_velocities[i].z += 0.5 * TF3A0;
+        translational_velocities[i].x += 0.5f * TF1A0;
+        translational_velocities[i].y += 0.5f * TF2A0;
+        translational_velocities[i].z += 0.5f * TF3A0;
 
-        rotational_velocities[i].x += 1.5 * (TF1A1 - (orientation_i.x * orientation_i.x * TF1A1 + orientation_i.x * orientation_i.y * TF2A1 + orientation_i.x * orientation_i.z * TF3A1));
-        rotational_velocities[i].y += 1.5 * (TF2A1 - (orientation_i.x * orientation_i.y * TF1A1 + orientation_i.y * orientation_i.y * TF2A1 + orientation_i.y * orientation_i.z * TF3A1));
-        rotational_velocities[i].z += 1.5 * (TF3A1 - (orientation_i.x * orientation_i.z * TF1A1 + orientation_i.y * orientation_i.z * TF2A1 + orientation_i.z * orientation_i.z * TF3A1));
+        rotational_velocities[i].x += 1.5f * (TF1A1 - (orientation_i.x * orientation_i.x * TF1A1 + orientation_i.x * orientation_i.y * TF2A1 + orientation_i.x * orientation_i.z * TF3A1));
+        rotational_velocities[i].y += 1.5f * (TF2A1 - (orientation_i.x * orientation_i.y * TF1A1 + orientation_i.y * orientation_i.y * TF2A1 + orientation_i.y * orientation_i.z * TF3A1));
+        rotational_velocities[i].z += 1.5f * (TF3A1 - (orientation_i.x * orientation_i.z * TF1A1 + orientation_i.y * orientation_i.z * TF2A1 + orientation_i.z * orientation_i.z * TF3A1));
     }
 
     translational_velocities[i].x /= d;
